* commit 'd7535eaf6836b3f5f6e522d4a29fe0e9b9e1583e': (21 commits)
  CC-1524 Remove the possibility of duplicate resolved languages
  CC-1524 Explain the possibility of duplicate resolved languages
  Update CrawlerDetect to latest release (v1.2.103)
  CC-2774 ConceptTree: delete topmostResources; use roots directly
  CC-2774 ConceptTree: iterate only over collections, where possible
  CC-1524 Tidy ParsedLanguage (remove surplus blank lines)
  CC-2522 Browse: fix/tidy commenting in ResourceRef
  Change PoolParty hostname used in mock server response
  Revert change to StatementHandler.java
  CC-1524 Update Registry schema doc: say BCP 47 instead of ISO 639-1
  CC-1524 Have parseLanguageTags return a list element for every input
  CC-1524 Update Registry schema changelog for new language elements
  CC-1524 Support language-list element in get vocabulary API methods
  CC-1524 Canonicalize language tags during vocabulary validation
  CC-1524 Have language tag parsing API methods return canonical form
  CC-1524 Replace language tag resolution with LSR implementation
  CC-1524 Add IANA LSR and properties for it
  Fix capitalization in comment in PropertyConstants
  Fix bogus chown command in INSTALL.md
  Lock down access to critical and private admin methods
  ...

+++ src/main/java/au/org/ands/vocabs/registry/db/utils/PopulateSubjectResolver.java
@@ -3,10 +3,14 @@
 package au.org.ands.vocabs.registry.db.utils;
 
 import java.lang.invoke.MethodHandles;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
+import javax.ws.rs.core.Context;
 
 import org.openrdf.OpenRDFException;
 import org.openrdf.model.Value;
@@ -28,6 +32,7 @@ import au.org.ands.vocabs.registry.db.entity.SubjectResolverSource;
 
 /** Utility class for populating the subject_resolver table based on
  * the contents of the subject_resolver_sources table.
+ * The populate method is accessible only via a loopback interface.
  */
 @Path(AdminApiPaths.API_ADMIN + "/" + AdminApiPaths.DATABASE)
 public class PopulateSubjectResolver {
@@ -50,12 +55,27 @@ public class PopulateSubjectResolver {
 
     /** Populate the subject_resolver table with the subjects extracted
      * from the sources given in the subject_resolver_sources table.
+     * @param request The HTTP request.
      * @return Text indicating success.
      */
     @Path("populateSubjectResolver")
     @GET
-    public String populateSubjectResolver() {
+    public String populateSubjectResolver(
+            @Context final HttpServletRequest request) {
         logger.info("In populateSubjectResolver");
+        // Only allow access from a loopback interface.
+        String remoteAddress = request.getRemoteAddr();
+        InetAddress ipAddress;
+        try {
+            ipAddress = InetAddress.getByName(remoteAddress);
+        } catch (UnknownHostException e) {
+            logger.error("Unable to parse IP address", e);
+            return "Unable to parse caller's IP address";
+        }
+
+        if (!ipAddress.isLoopbackAddress()) {
+            return "Not localhost; forbidden";
+        }
 
         // First, clean out any existing entries.
         SubjectResolverEntryDAO.deleteAllResolverEntries();
